+++ core/src/evm/interpreter/mod.rs
@@ -51,7 +51,9 @@ use crate::{
     },
 };
 use bit_set::BitSet;
-use cfx_types::{Address, BigEndianHash, H256, U256, U512};
+use cfx_types::{
+    address_util::AddressUtil, Address, BigEndianHash, H256, U256, U512,
+};
 use std::{cmp, convert::TryFrom, marker::PhantomData, mem, sync::Arc};
 
 const GASOMETER_PROOF: &str = "If gasometer is None, Err is immediately returned in step; this function is only called by step; qed";
@@ -898,8 +900,11 @@ impl<Cost: CostType> Interpreter<Cost> {
                 // clear return data buffer before creating new call frame.
                 self.return_data = ReturnData::empty();
 
-                let can_call =
-                    has_balance && context.depth() < context.spec().max_depth;
+                let valid_code_address = code_address.is_valid_address();
+
+                let can_call = has_balance
+                    && context.depth() < context.spec().max_depth
+                    && valid_code_address;
                 if !can_call {
                     self.stack.push(U256::zero());
                     return Ok(InstructionResult::UnusedGas(call_gas));
@@ -984,7 +989,8 @@ impl<Cost: CostType> Interpreter<Cost> {
             }
             instructions::SUICIDE => {
                 let address = self.stack.pop_back();
-                context.suicide(&u256_to_address(&address))?;
+                let refund_address = u256_to_address(&address);
+                context.suicide(&refund_address)?;
                 return Ok(InstructionResult::StopExecution);
             }
             instructions::LOG0
@@ -1629,7 +1635,7 @@ mod tests {
 
     #[test]
     fn should_not_fail_on_tracing_mem() {
-        let code = "7feeffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff006000527faaffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffaa6020526000620f120660406000601773945304eb96065b2a98b57a48a06ae28d285a71b56101f4f1600055".from_hex().unwrap();
+        let code = "7feeffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff006000527faaffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffaa6020526000620f120660406000601773145304eb96065b2a98b57a48a06ae28d285a71b56101f4f1600055".from_hex().unwrap();
 
         let mut params = ActionParams::default();
         params.address = Address::from_low_u64_be(5);